Prevented from missing base policy in foreman-selinux
We've hit issue downstream when base policy was not loaded in the buildroot
causing some optional_blocks to be ignored. We should error out early when
base policy is not loaded in the build root to prevent that, otherwise we
will likely find this in user reported bugs (allow rule is missing), because
SELinux will silently ignore those blocks.