Add version of Foreman this is built against to release
Building on rpm/develop (1.8) introduces changes incompatible with rpm/1.7, so
the package needs building on both.  Since rubygem-foreman_docker-1.2.0-1 can't
be built twice (duplicate NEVRA will cause errors in both tito tagging and in
Koji), it needs to differ.

Taking inspiration from EL's z-stream builds (.el6_6 in a dist tag), inserting
the Foreman version that this is being built for will allow us to have a
different NEVRA.

In future, we could perhaps define a %{foreman_dist} for use in these tags.