Partially revert "Better branched builds with Foreman version macro"
This reverts commit f1375ed812dbcc51cb9a9a4b0926c72ce80134a1.

This solution isn't compatible with koji, as the SRPM generated and submitted
by tito doesn't contain the expanded foremandist tag.  Koji uses the EVR from
the SRPM as the build ID and then fails the build as the EVR in the binary RPM
doesn't match.

Once dgoodwin/tito#155 is fixed, we can configure the releasers to define
foremandist so it's expanded in the submitted SRPMs.

(Koji's support for building SRPMs from SCM may also work, as it would use the
srpm-build group under mock, which can expand foremandist.)