Replace plugin upgrade/install test to fix Bundler errors
When a plugin is upgraded and the version changes, the `bundle show`
command always failed as bundler.d/ had already been updated, but the
Gemfile.lock and vendor/ area was out of date.

Error from an upgrade where 1.0.1 was currently installed, but
bundler.d/ had been unpacked and stated version 1.0.1:

    # bundle show deface
    Could not find deface-1.0.2 in any of the sources
    Bundler::GemNotFound: Could not find deface-1.0.2 in any of the
    sources

This caused the postinst to always call `bundle install`, even on
upgrades.  For plugins that are dependencies of others, this causes an
error as `bundle update` should be used.

Checking the arguments passed into the postinst by dpkg to check for an
upgrade is more reliable.