fixes #12912 - switch to CentOS SCLo SIG builds
Swaps dependencies for softwarecollections.org with the CentOS SCLo SIG
release packages, which are copied into our repo by the mash script when
listed in extras/ so non-CentOS users and those without CentOS Extras
enabled can install it.

The packages are better maintained and will allow us to depend on sclo-*
packages in future.  Maintenance of softwarecollections.org builds seems
to have stopped in favour of the SCLo SIG.