Link foreman-ruby to specific ruby version
When building a plugin that has a build-dep on `foreman`, or just
installing the package, the foreman-ruby symlink used in
foreman.postinst will be non-functional as it points to /usr/bin/ruby on
on jessie/xenial. This only exists when the `ruby` package is installed,
but `foreman` depends on a specific `rubyX.Y` package.