fixes #17134 - fix release bump upgrade to keep gem installed
The package preun ran after the post scriptlet, which when the release
and _not_ version was updated, caused the gem to be uninstalled as both
the old and new packages managed the same version.

The gem is now only uninstalled in preun if there are multiple versions
installed (i.e. a second version from the new package). A workaround was
added for the current version 0.5.1 to restore the gem on upgrade.