Make foreman-selinux depend on container-selinux
foreman-selinux used to define a type docker_port_t because there was no
official policy that defined this type.

container-selinux has this type too, and both policies cannot be
installed together. Since container-selinux is the default policy on EL
systems for container stuff, foreman-selinux shouldn't stop it from
being installed, and depend on it instead for the types it needs.