Fixes #22181 - Change katello-change-hostname work with custom certs
katello-change-hostname fails with custom certs currently. Currently
there are no arguments to pass new certs that match the new hostname
to the installer. This commit adds those options and some checks
to see if the user is using custom certs. If they are using custom
certs, katello-certs-check will run before any changes are made.
On proxies w/ content, a certs tar must be made with the new
custom certs. This is passed into the existing --certs-tar option.