Move foreman-rails repo to foreman-release
This moves the foreman-rails repo to foreman-release. It's tied to the
foreman release cycle and is versioned in the same way. This avoids a
lot of the problems with ensuring we have the proper version installed
and obsolete the older versions.

We also only depend on centos-release-scl-rh. We used to require SCLo
repositories for Rails 4.2 but now we ship our own repository for this.