Fix foreman NPM deps and rails GPG key
The GPG key was in an incorrect format which made it useless.

The NPM dependencies are synced with package.json. This also fixes a
dependency that was in foreman on a npm package. It is now in
foreman-assets which means we don't need to install and node packages on
production systems.