Fixes #24928 - generators for requires/provides of webpack bundles
This adds automatic Provides and Requires for Foreman's webpack bundles.

Foreman ships webpack bundles in its RPMs to be consumed by plugins. To ensure
proper dependency relationships between the plugins and the main Foreman RPM,
this script generates Provides for the Foreman RPM that contain the hash of
the webpack bundles shipped and matching Requires for the plugins packages.
Each time the Foreman bundles change hashes, the plugins become uninstallable
and indicate that a rebuild is required.