Remove webpack provides/requires from Foreman
The goal of this was to make sure we always built webpack against a
bundle that was compatible, or were forced to rebuilt it. That part sort
of worked, but the idea was that a rebuild with an incompatible build
would fail.

The reality is that webpack will happy use any dependency, even if it
doesn't fit the interface. This can be seen by various failures where a
webpack bundle was generated that was utterly broken.

Since it doesn't really provide the guarantees we wanted but does add a
lot of overhead, it's better to remove this.