Add a script to compare package versions to bundler
This has a script build_cache to build a cache with all versions with
the following structure:

{
  "tfm-rubygem-my-gem": {
    "version": "1.2.3",
    "spec_file": "packages/foreman/rubygem-my-gem/rubygem-my-gem.spec"
  }
}

The compare-gems script is passed a Gemfile to ask bundler to resolve
its dependencies. Then it calculates the RPM package names. Since
build_cache uses a cache file, it can always call the script and compare
the bundler versions to the packaged versions.