Fixes #29424 - %check section for gems with native libs
Since we manually move file to their correct locations, it happens that
they end up in the wrong place. The RPM build then passes but the
resulting package is broken. By adding a %check section that requires
the gem, this should be prevented in most cases.

The EL8 check fails for some unknown reason and needs to be looked into.