Refs #29817 - support Dynflow worker pool
This reflects a change in the installer where a dynflow pool is
supported. This means workers are now named worker-N, starting at 1. To
prevent the Foreman RPM from placing these files again, the change is
reflected here.

Another change is that the dynflow config files are now owned by
foreman-dynflow-sidekiq instead of the main foreman package.