foreman-bootloaders-redhat: Set prefix "grub2" in generated Grub2.
This is the actual prefix used in the tftpboot layout,
and embedding it into the image increases compatibility.

An example case are DHCP servers sending the boot_file in
DHCP option 67 instead of using the classic BOOTP field
to conserve space.
This is not yet handled in Grub2 < 2.0.4, which then fails
to find its configuration unless the prefix is known.