Fixes #32003 - Require Puppet with DNF module support
On EL8 DNF module support in Puppet is needed. This was introduced in
Puppet 5.5.20 and Puppet 6.15.0. By requiring this at the packaging
level, any incompatibility is handled. It uses rich dependencies only
available on EL8.

This is only relevant for the katello package so it's added to the
foreman-installer-katello package.