Set module_hotfixes to true for repoclosure on EL8
The move to Ruby 2.7 means needing a module that is not the default
profile. There is no way with repoclosure to set which module stream
to use when computing repoclosure. Adding module_hotfixes to the
AppStream repository definition for repoclosure allows repoclosure
to look at the AppStream repository as a regular/flat repository
and resolve dependencies.