Update package testing instructions
This is what I used locally. The plugins and Katello side are not
included since I didn't test those.

Note that I also needed a newer obal for mock (or really, srpm) to work.

Another thing to note is that this no longer automatically determines
the git prerelease sha which the tito config did.