Refs #33406: Move scenarios to /var/lib/foreman-installer
Supports the move of scenario answers and config to /var/lib (sharedstatedir)
and out of system config directory. On upgrade, before installing the
new version of the package the scenario files are moved out of /etc
and into /var/lib to allow migrations to apply on the new location properly.
The custom-hiera.yaml remains in /etc as users are expected to edit this file.