Stop calling gem compare in bump_rpm.sh
This command was very useful to detect new dependency updates, but most
gems now have code where update-requirements can actually update it
automatically. That means it's mostly useless. The file changes are
generally not that interesting. In addition to that, it's also broken on
Ruby 3 and has old outdated dependencies.