Add a script to list updatable packages
This script lists all directories which can be updated using bump_rpm.sh
and done so safely in an automated manner, including the version number.
Nightly packages are filtered out since they can't be bumped.

If GitHub actions are detected it also outputs the data in a way that
integrates. Then the output can be registered and used in other steps.