rename stream to el8 to avoid problems with puppet
for puppet `latest` has a special meaning, so if you instruct it to
enable the `foreman` module, `latest` stream, it actually looks for the
latest version available, not for the stream with the literal name
`latest`.

looking at the output of `dnf module list` on an EL8 machine, most
streams use the version of the software here, or `rhel8`.
we didn't want to use the version, as that would mean our users need to
explicitly `dnf module reset` and `dnf module enable` the newer version
on upgrades, which seems like a burden given the fact that our repo will
never contain multiple release streams at once anyways.

as Foreman also works fine on non-RH EL8, I propose the name `el8`

possible alternatives: current, released?