Require python38-psutil on EL8 when using ansible-core
ansible-core uses Python 3.8 (vs 3.6 as used by ansible 2.9), and to be
able to properly execute that, we need the psutil module, as otherwise
ansible fails to execute.

This is not a packaging issue with ansible-core itself, as that doesn't
require psutil, but with ansible-runner. However, our chances to get a
fixed runner RPM anytime soon are rather limited.