Enable building directly through Koji CLI
This drops the need for use of tito to handle building scratch or
release builds of RPMs and uses Koji directly through the CLI in Obal
commands. This will allow dropping the tito functionality and relying
on the package manifest as the source for all configuration and
for doing invidualized builds based on tag. This also paves the way
to make tagging of existing builds a part of the workflow.