Use modularity in repoclosure
This removes the module_hotfixes and excludepkgs hack. To use a
different modules.d directory, the --installroot parameter must be
passed. This in turn requires the releasever to be set. This is actually
also useful in determining which module(s) need to be enabled. Passing
an installroot also uses a correct cache directory and further isolates
dnf from the host system.

Additionally a template is passed to mktemp for easier debugging.