Turn on auto-requires for rubygem-nokogiri
In the new packaging template we rely on auto-requires to generate the
runtime dependencies. Turning this off but not stating the dependencies
breaks at runtime.

It hacks around it by removing racc as a dependency. The explicit
versions are stated so the build fails if they are ever changed. This
means they shouldn't go out of sync.

Fixes: d521dd04e9f4bd0b132aad86935c1fb9b6ce81bb