Showcase dynamic build requires
Since Fedora 31 it's possible to use Dynamic BuildRequires[1]. This
removes the need to change the spec file all the time to keep in sync
with the Foreman repo itself.

Once EL 8 support is dropped, the legacy way can be dropped. Until then
it's hidden behind conditionals so at least EL 9 is up to date.

In a proper implementation we would have tooling that's usable in all
package, like a foreman-packaging RPM with reusable scripts and macros.

[1]: https://fedoraproject.org/wiki/Changes/DynamicBuildRequires