Showcase dynamic build requires
Since Fedora 31 and EL9 it's possible to use Dynamic BuildRequires.
This removes the need to change the spec file all the time to keep in
sync with the Foreman repo itself.

In a proper implementation we would have tooling that's usable in all
package, like a foreman-packaging RPM with reusable scripts and macros.

Link: https://fedoraproject.org/wiki/Changes/DynamicBuildRequires