Fixes #35461 - Require /usr/sbin/sendmail to be available
Foreman uses sendmail as the default mail handler, but modern systems
might not have it available by default.

Only depend on the path, not an explicit package, as there might be
different packages providing a sendmail binary (e.g. on EL8 there are at
least Postfix and Sendmail doing so).