Add rubygem-webrick package
This package is part of Ruby 2.7's default gems, but in Ruby 3.1 it's no
longer there. That's why packages have started to add this dependency.
Since ruby-default-gems on EL8 doesn't have Provides: rubygem(webrick)
it isn't satisfied. Rather than hacking around it in every gem that
requires it, this starts to package webrick explicitly. In the future it
needs to happen anyway.