Fixes #35743 - Introduce foreman-obsolete-packages
Sometimes packages are removed from our repositories and we want them to
be cleaned up automatically. In case there is no replacement (like when
a Foreman plugin is retired), there is nothing to obsolete it.

On EL7 we could abuse the tfm package, but no such thing exists on EL8.
Fedora has fedora-obsolete-packages for this purpose.
foreman-obsolete-packages should fulfill the same purpose.