Ensure tmp directory exists in foreman precompile plugin
This was first discovered as a problem during some Katello on EL8
packaging pipelines which failed to run Katello due to the missing
directory. Since then that particular type of Katello run has been
adjusted to be independent of the existence of the tmp directory.
However, also recreating the directory here, makes this macro more
robust and internally consistent.