Add a foreman-discovery-image package
While this could be built in Koji, this actually takes an easier
approach and downloads the image from downloads.theforeman.org. It
should be noted that the paths there aren't easy to consume, since it
includes a git hash. Possibly that's a bug in our release pipeline and
it may be easier to fix that than to keep the complex part here.