Depend on libopenscap.so.25 explicitly
In the past we depended on the package name, but that was to support
both EL7 & EL8. Now we're EL8 only so we can make it more narrow. That
means we no longer need to maintain a version range, which breaks on
some EL updates.

It copies the library dependency bit from rubygems-ruby-vips.spec, so it
correctly works for all arches as was recommended on the Fedora devel
list[1].

[1]: https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/