Make katello-client-bootstrap files owned by root
There was a report from a user who saw errors in dnf.log:

> 2023-01-20T16:20:05+0200 SUBDEBUG Installed:
> katello-client-bootstrap-1.7.9-1.el8sat.noarch
> 2023-01-20T16:20:05+0200 INFO warning: user apache does not exist - using
> root
> warning: group apache does not exist - using root
> warning: user apache does not exist - using root
> warning: group apache does not exist - using root

That implies the package was installed before httpd-filesystem (which
normally creates the user + group), or something else was wrong. It's
weird because it should depend on /var/www/html provided by
httpd-filesystem.

This changes the files to be owned by root, which means apache can't
modify them. This is actually desired. Since they're world readable, it
should not be a problem and may solve an edge case.