include foreman_assets_* in files also when webpack only
for both, sprockets and webpack based, assets we use
plugin:assets:precompile, which generates the content under
foreman_assets_* regardless whether sprockets assets are present or not.

for plugins that only ship webpack based assets, this means that the
template will generate a file list that is insufficient and rpmbuild
will complain about unpackaged files. not sure those files are really
needed as we don't have sprockets in that case, but its easier to just
include them for now.