Implement rich dependencies in gem2rpm templates
For most templates this is pretty straight forward, but for
foreman_plugin.spec.erb it combines asset reequirements into the
spec.extensions block.  That was already using the more modern
runtime_requirements.virtualize methods to generate a correct rpm
dependency list. This was never noticed because we (thankfully) don't
have foreman plugins with native extensions.