Reimplement repoquery in a more generic way
This approach is based on rpmdist-repoquery where a certain directory
structure is used to track repositories. That is then selected, together
with a releasever.

Our own repositories are made generic by using DNF_VAR_$var, which makes
branching easier.

It also uses true modularity. For that it hacks together a temporary
installroot and uses fakeroot to pretend to be root, because dnf insists
you need to be root to enable a module. Due to only using real DNF, we
can rely on the modulary metadata and dependency resolution to select
the correct modules. Again, reducing duplication.

A major motivation for this is the upcoming CentOS Stream 8 EOL where
the repositories will be archived. This new repository structure allows
easy switching to another Enterprise Linux for repoclosure checks.

This needs a different approach in how we configure repoclosure in obal
and the wrapper scripts should probably live in another repository as
well. Another excercise is the EL7 repoclosure check.