Include GPG keys in katello-repos and check for non-nightly
Including the GPG keys is better because it gives a higher degree of
trust because the RPM is GPG signed. It also helps because dnf doesn't
refresh HTTPS URLs and purging old entries is tedious. While the repos
aren't typically installed that long, I've had it happen to me with one
of my personal COPR repositories.

The downside is that after branching you need to update the GPG key in
the repository.

It also makes GPG key checking correct by looking at the correct values.
This allows enabling GPG for candlepin (which is using a signed release)
while not enabling it for Pulp or Katello (which are nightly).