Add foreman_ygg_migration package
This package should get pulled in when yggdrasil >=0.4.z is installed.

When installed, it checks whether yggdrasild.service is enabled. If not, it
does nothing. If enabled, this package tweaks yggdrasil's config file to account
for the newly needed config options and migrates service overrides from the old
service to the new worker. Additionally, it disables the yggdrasild service and
enables yggdrasil.service.

Our worker is left to be activated through dbus.